tests: extract test_masters_base.py and make the masters suite consistent
The masters tests were split by module, but the shared abstract engine
(BaseMasterModule in modules/masters/base.py) had no home -- its unit tests
were buried in test_master_dark.py, and the shared L1-output tests were
scattered (dtype provenance in bias; save_master duplicated asymmetrically
across bias and dark). Reorganize by *what is exercised*:

- test_masters_base.py (new): the shared engine, driven through the simplest
  concrete vehicle per path -- Dark (bias-subtracted) for calibration
  resolve/apply/load + frame-load guards + stacking (rate estimator, per-pixel
  rejection, datacube clipping, validation) + array cleaning; Bias (no
  calibrations) for the shared L1 dtype-provenance and save_master paths.
- test_master_bias.py / test_master_dark.py: now symmetric mirrors covering
  only each module's own behavior (Unit / Info / RoundTrip / Signature /
  Regression).
- save_master is a base method with no per-module override, so its tests
  (previously 8 across bias+dark, via two vehicles) collapse to one
  TestSaveMaster (5, via Bias); no unique behavior lost. Also fixes the stale
  `master_path` test naming (the kwarg is `filepath`).
- test_master_wls.py unchanged: WLS builds an ML2 and does not use the L1
  stacking engine, so it stays tested per-WLS-method on its own.

Docs: add a "Masters test layout" note to CLAUDE.md stating the belongs-in-base
rule; clarify the profiling paragraph (profiling folds base.py into bias/dark;
tests isolate it) so the two no longer appear to contradict. _masters.py
docstring updated (now shared by the base suite too).
